+++ core2/bindings.rs
@@ -12,7 +12,6 @@ use rusty_v8 as v8;
 use std::cell::Cell;
 use std::convert::TryFrom;
 use std::io::{stdout, Write};
-use std::option::Option;
 use url::Url;
 use v8::MapFnTo;
 
@@ -168,39 +167,6 @@ pub fn boxed_slice_to_uint8array<'sc>(
   let ab = v8::ArrayBuffer::with_backing_store(scope, &backing_store_shared);
   v8::Uint8Array::new(scope, ab, 0, buf_len)
     .expect("Failed to create UintArray8")
-}
-
-pub extern "C" fn host_import_module_dynamically_callback(
-  _context: v8::Local<v8::Context>,
-  _referrer: v8::Local<v8::ScriptOrModule>,
-  _specifier: v8::Local<v8::String>,
-  _import_assertions: v8::Local<v8::FixedArray>,
-) -> *mut v8::Promise {
-  todo!()
-}
-
-pub extern "C" fn host_initialize_import_meta_object_callback(
-  context: v8::Local<v8::Context>,
-  module: v8::Local<v8::Module>,
-  meta: v8::Local<v8::Object>,
-) {
-  let scope = &mut unsafe { v8::CallbackScope::new(context) };
-  let state_rc = JsRuntime::state(scope);
-  let state = state_rc.borrow();
-
-  let module_global = v8::Global::new(scope, module);
-  let info = state
-    .module_map
-    .get_info(&module_global)
-    .expect("Module not found");
-
-  let url_key = v8::String::new(scope, "url").unwrap();
-  let url_val = v8::String::new(scope, &info.name).unwrap();
-  meta.create_data_property(scope, url_key.into(), url_val.into());
-
-  let main_key = v8::String::new(scope, "main").unwrap();
-  let main_val = v8::Boolean::new(scope, info.main);
-  meta.create_data_property(scope, main_key.into(), main_val.into());
 }
 
 pub extern "C" fn promise_reject_callback(message: v8::PromiseRejectMessage) {
@@ -722,47 +688,6 @@ fn shared_getter(
   rv.set(shared_ab.into())
 }
 
-// Called by V8 during `Isolate::mod_instantiate`.
-pub fn module_resolve_callback<'s>(
-  context: v8::Local<'s, v8::Context>,
-  specifier: v8::Local<'s, v8::String>,
-  _import_assertions: v8::Local<'s, v8::FixedArray>,
-  referrer: v8::Local<'s, v8::Module>,
-) -> Option<v8::Local<'s, v8::Module>> {
-  let scope = &mut unsafe { v8::CallbackScope::new(context) };
-
-  let state_rc = JsRuntime::state(scope);
-  let state = state_rc.borrow();
-
-  let referrer_global = v8::Global::new(scope, referrer);
-  let referrer_info = state
-    .module_map
-    .get_info(&referrer_global)
-    .expect("ModuleInfo not found");
-  let referrer_name = referrer_info.name.to_string();
-
-  let specifier_str = specifier.to_rust_string_lossy(scope);
-
-  use crate::ModuleSpecifier;
-  // FIXME(bartlomieju): import map support
-  let resolved_specifier =
-    ModuleSpecifier::resolve_import(&specifier_str, &referrer_name)
-      .expect("Module should have been already resolved");
-
-  if let Some(id) = state.module_map.get_id(resolved_specifier.as_str()) {
-    if let Some(handle) = state.module_map.get_handle(id) {
-      return Some(v8::Local::new(scope, handle));
-    }
-  }
-
-  let msg = format!(
-    r#"Cannot resolve module "{}" from "{}""#,
-    specifier_str, referrer_name
-  );
-  throw_type_error(scope, msg);
-  None
-}
-
 // Returns promise details or throw TypeError, if argument passed isn't a Promise.
 // Promise details is a js_two elements array.
 // promise_details = [State, Result]
@@ -853,7 +778,7 @@ fn get_proxy_details(
   rv.set(proxy_details.into());
 }
 
-fn throw_type_error(scope: &mut v8::HandleScope, message: impl AsRef<str>) {
+pub fn throw_type_error(scope: &mut v8::HandleScope, message: impl AsRef<str>) {
   let message = v8::String::new(scope, message.as_ref()).unwrap();
   let exception = v8::Exception::type_error(scope, message);
   scope.throw_exception(exception);
