+++ core2/lib.rs
@@ -44,13 +44,8 @@ pub use crate::async_cell::RcRef;
 pub use crate::flags::v8_set_flags;
 pub use crate::module_specifier::ModuleResolutionError;
 pub use crate::module_specifier::ModuleSpecifier;
-pub use crate::modules::FsModuleLoader;
 pub use crate::modules::ModuleId;
-pub use crate::modules::ModuleLoadId;
-pub use crate::modules::ModuleLoader;
 pub use crate::modules::ModuleSource;
-pub use crate::modules::ModuleSourceFuture;
-pub use crate::modules::NoopModuleLoader;
 pub use crate::normalize_path::normalize_path;
 pub use crate::ops::json_op_async;
 pub use crate::ops::json_op_sync;
