+++ core2/runtime.rs
@@ -3,14 +3,10 @@
 use rusty_v8 as v8;
 
 use crate::bindings;
-use crate::error::attach_handle_to_error;
 use crate::error::generic_error;
 use crate::error::AnyError;
-use crate::error::ErrWithV8Handle;
 use crate::error::JsError;
-use crate::modules::ModuleId;
 use crate::modules::ModuleMap;
-use crate::modules::ModuleSource;
 use crate::ops::*;
 use crate::shared_queue::SharedQueue;
 use crate::shared_queue::RECOMMENDED_SIZE;
@@ -25,7 +21,6 @@ use std::any::Any;
 use std::cell::Cell;
 use std::cell::RefCell;
 use std::collections::HashMap;
-use std::convert::TryFrom;
 use std::ffi::c_void;
 use std::mem::forget;
 use std::option::Option;
@@ -282,10 +277,10 @@ impl JsRuntime {
     isolate.set_capture_stack_trace_for_uncaught_exceptions(true, 10);
     isolate.set_promise_reject_callback(bindings::promise_reject_callback);
     isolate.set_host_initialize_import_meta_object_callback(
-      bindings::host_initialize_import_meta_object_callback,
+      crate::modules::host_initialize_import_meta_object_callback,
     );
     isolate.set_host_import_module_dynamically_callback(
-      bindings::host_import_module_dynamically_callback,
+      crate::modules::host_import_module_dynamically_callback,
     );
     isolate
   }
@@ -550,247 +545,6 @@ pub(crate) fn exception_to_err_result<'s, T>(
 
 // Related to module loading
 impl JsRuntime {
-  fn compile_module(
-    &mut self,
-    specifier: &str,
-    source: &str,
-  ) -> Result<v8::Global<v8::Module>, AnyError> {
-    let context = self.global_context();
-    let scope = &mut v8::HandleScope::with_context(self.v8_isolate(), context);
-
-    let specifier_str = v8::String::new(scope, specifier).unwrap();
-    let source_str = v8::String::new(scope, source).unwrap();
-
-    let origin = bindings::module_origin(scope, specifier_str);
-    let source = v8::script_compiler::Source::new(source_str, &origin);
-
-    let tc_scope = &mut v8::TryCatch::new(scope);
-
-    let maybe_module = v8::script_compiler::compile_module(tc_scope, source);
-
-    if tc_scope.has_caught() {
-      assert!(maybe_module.is_none());
-      let e = tc_scope.exception().unwrap();
-      return exception_to_err_result(tc_scope, e, false);
-    }
-
-    let module = maybe_module.unwrap();
-    let module_handle = v8::Global::<v8::Module>::new(tc_scope, module);
-    Ok(module_handle)
-  }
-
-  // TODO(bartlomieju): this can be a method on the `ModuleMap`
-  /// Low-level module creation.
-  ///
-  /// Called during module loading or dynamic import loading.
-  pub fn create_module(
-    &mut self,
-    info: ModuleSource,
-    main: bool,
-  ) -> Result<ModuleId, AnyError> {
-    let state_rc = Self::state(self.v8_isolate());
-
-    if info.module_url_specified != info.module_url_found {
-      state_rc
-        .borrow_mut()
-        .module_map
-        .alias(&info.module_url_specified, &info.module_url_found);
-    }
-
-    let maybe_module_id =
-      state_rc.borrow().module_map.get_id(&info.module_url_found);
-
-    if let Some(id) = maybe_module_id {
-      // Module has already been registered.
-      debug!(
-        "Already-registered module fetched again: {}",
-        info.module_url_found
-      );
-      return Ok(id);
-    }
-
-    let module_handle =
-      self.compile_module(&info.module_url_found, &info.code)?;
-    let id = state_rc.borrow_mut().module_map.register(
-      &info.module_url_found,
-      main,
-      module_handle,
-    );
-
-    Ok(id)
-  }
-
-  /// Instantiates a ES module
-  ///
-  /// `AnyError` can be downcast to a type that exposes additional information
-  /// about the V8 exception. By default this type is `JsError`, however it may
-  /// be a different type if `RuntimeOptions::js_error_create_fn` has been set.
-  pub fn mod_instantiate(&mut self, id: ModuleId) -> Result<(), AnyError> {
-    let state_rc = Self::state(self.v8_isolate());
-    let context = self.global_context();
-
-    let scope = &mut v8::HandleScope::with_context(self.v8_isolate(), context);
-    let tc_scope = &mut v8::TryCatch::new(scope);
-
-    let module = state_rc
-      .borrow()
-      .module_map
-      .get_handle(id)
-      .map(|handle| v8::Local::new(tc_scope, handle))
-      .expect("ModuleInfo not found");
-
-    if module.get_status() == v8::ModuleStatus::Errored {
-      exception_to_err_result(tc_scope, module.get_exception(), false)?
-    }
-
-    let result =
-      module.instantiate_module(tc_scope, bindings::module_resolve_callback);
-    match result {
-      Some(_) => Ok(()),
-      None => {
-        let exception = tc_scope.exception().unwrap();
-        exception_to_err_result(tc_scope, exception, false)
-      }
-    }
-  }
-
-  pub async fn mod_evaluate(&mut self, id: ModuleId) -> Result<(), AnyError> {
-    let state_rc = Self::state(self.v8_isolate());
-
-    let maybe_promise_handle = {
-      let context = self.global_context();
-      let scope =
-        &mut v8::HandleScope::with_context(self.v8_isolate(), context);
-
-      let module = state_rc
-        .borrow()
-        .module_map
-        .get_handle(id)
-        .map(|handle| v8::Local::new(scope, handle))
-        .expect("ModuleInfo not found");
-      let mut status = module.get_status();
-      assert_eq!(status, v8::ModuleStatus::Instantiated);
-
-      // IMPORTANT: Top-level-await is enabled, which means that return value
-      // of module evaluation is a promise.
-      //
-      // Because that promise is created internally by V8, when error occurs during
-      // module evaluation the promise is rejected, and since the promise has no rejection
-      // handler it will result in call to `bindings::promise_reject_callback` adding
-      // the promise to pending promise rejection table - meaning JsRuntime will return
-      // error on next poll().
-      //
-      // This situation is not desirable as we want to manually return error at the
-      // end of this function to handle it further. It means we need to manually
-      // remove this promise from pending promise rejection table.
-      //
-      // For more details see:
-      // https://github.com/denoland/deno/issues/4908
-      // https://v8.dev/features/top-level-await#module-execution-order
-      let maybe_value = module.evaluate(scope);
-
-      // Update status after evaluating.
-      status = module.get_status();
-
-      if let Some(value) = maybe_value {
-        assert!(
-          status == v8::ModuleStatus::Evaluated
-            || status == v8::ModuleStatus::Errored
-        );
-        let promise = v8::Local::<v8::Promise>::try_from(value)
-          .expect("Expected to get promise as module evaluation result");
-        let promise_global = v8::Global::new(scope, promise);
-        // FIXME(bartlomieju): comment above
-        state_rc
-          .borrow_mut()
-          .pending_promise_exceptions
-          .remove(&promise_global);
-        scope.perform_microtask_checkpoint();
-        Some(promise_global)
-      } else {
-        // FIXME(bartlomieju): this path depends on the comment above and
-        // promise rejection being added to `state.pending_promise_expcetions`
-        assert!(status == v8::ModuleStatus::Errored);
-        None
-      }
-    };
-
-    // FIXME(bartlomieju): this path depends on the comment above and
-    // promise rejection being added to `state.pending_promise_expcetions`
-    if maybe_promise_handle.is_none() {
-      let err = self.check_promise_exceptions().unwrap_err();
-      return Err(err);
-    }
-
-    let promise_handle = maybe_promise_handle.unwrap();
-
-    poll_fn(|cx| {
-      let _r = self.poll_event_loop(cx)?;
-
-      // Top level module
-      let maybe_result = self.evaluate_pending_module(promise_handle.clone());
-
-      if let Some(result) = maybe_result {
-        // TODO(bartlomieju): is it ok?
-        return Poll::Ready(result);
-      }
-
-      let state = state_rc.borrow();
-      if state.pending_ops.is_empty() {
-        let msg = "Module evaluation is still pending but there are no pending ops or dynamic imports. This situation is often caused by unresolved promise.";
-        return Poll::Ready(Err(generic_error(msg)));
-      }
-
-      Poll::Pending
-    })
-    .await
-  }
-
-  // TODO(bartlomieju): rename me
-  /// "deno_core" runs V8 with "--harmony-top-level-await"
-  /// flag on - it means that each module evaluation returns a promise
-  /// from V8.
-  ///
-  /// This promise resolves after all dependent modules have also
-  /// resolved. Each dependent module may perform calls to "import()" and APIs
-  /// using async ops will add futures to the runtime's event loop.
-  /// It means that the promise returned from module evaluation will
-  /// resolve only after all futures in the event loop are done.
-  ///
-  /// Thus during turn of event loop we need to check if V8 has
-  /// resolved or rejected the promise. If the promise is still pending
-  /// then another turn of event loop must be performed.
-  fn evaluate_pending_module(
-    &mut self,
-    promise_handle: v8::Global<v8::Promise>,
-  ) -> Option<Result<(), AnyError>> {
-    let context = self.global_context();
-    let scope = &mut v8::HandleScope::with_context(self.v8_isolate(), context);
-
-    let promise = promise_handle.get(scope);
-    let promise_state = promise.state();
-
-    match promise_state {
-      v8::PromiseState::Pending => {
-        // pass, poll_event_loop will decide if
-        // runtime would be woken soon
-        None
-      }
-      v8::PromiseState::Fulfilled => {
-        scope.perform_microtask_checkpoint();
-        Some(Ok(()))
-      }
-      v8::PromiseState::Rejected => {
-        let exception = promise.result(scope);
-        scope.perform_microtask_checkpoint();
-        let err1 = exception_to_err_result::<()>(scope, exception, false)
-          .map_err(|err| attach_handle_to_error(scope, err, exception))
-          .unwrap_err();
-        Some(Err(err1))
-      }
-    }
-  }
-
   fn poll_pending_ops(
     &mut self,
     cx: &mut Context,
@@ -843,7 +597,7 @@ impl JsRuntime {
     overflow_response
   }
 
-  fn check_promise_exceptions(&mut self) -> Result<(), AnyError> {
+  pub fn check_promise_exceptions(&mut self) -> Result<(), AnyError> {
     let state_rc = Self::state(self.v8_isolate());
     let mut state = state_rc.borrow_mut();
 
@@ -959,6 +713,7 @@ impl JsRuntime {
 #[cfg(test)]
 pub mod tests {
   use super::*;
+  use crate::modules::ModuleSource;
   use crate::modules::*;
   use crate::BufVec;
   use futures::future::lazy;
@@ -1687,7 +1442,8 @@ pub mod tests {
       "#
         .to_string(),
       };
-      let mod_a = runtime.create_module(info, true).unwrap();
+      let mod_a =
+        crate::modules::create_module(&mut runtime, info, true).unwrap();
       assert_eq!(dispatch_count.load(Ordering::Relaxed), 0);
 
       let info = ModuleSource {
@@ -1695,15 +1451,17 @@ pub mod tests {
         module_url_found: "file:///b.js".to_string(),
         code: "export function b() { return 'b' }".to_string(),
       };
-      let mod_b = runtime.create_module(info, false).unwrap();
+      let mod_b =
+        crate::modules::create_module(&mut runtime, info, false).unwrap();
 
-      runtime.mod_instantiate(mod_b).unwrap();
+      mod_instantiate(&mut runtime, mod_b).unwrap();
       assert_eq!(dispatch_count.load(Ordering::Relaxed), 0);
 
-      runtime.mod_instantiate(mod_a).unwrap();
+      mod_instantiate(&mut runtime, mod_a).unwrap();
       assert_eq!(dispatch_count.load(Ordering::Relaxed), 0);
 
-      let mut mod_evaluate_future = runtime.mod_evaluate(mod_a).boxed_local();
+      let mut mod_evaluate_future =
+        crate::modules::mod_evaluate(&mut runtime, mod_a).boxed_local();
 
       let _result = mod_evaluate_future.poll_unpin(cx);
       assert_eq!(dispatch_count.load(Ordering::Relaxed), 1);
